ai_service/requirements.txt
fastapi>=0.135.0
uvicorn[standard]>=0.34.0
google-genai>=1.0.0
pydantic>=2.12.0
pydantic-settings>=2.7.0
redis>=5.2.0
python-dotenv>=1.0.0
httpx>=0.28.0
tenacity>=9.0.0
